Move test into core
