+++ core/data/src/test/java/ru/d3rvich/core/data/ExampleUnitTest.kt
@@ -1,7 +1,8 @@
-package ru.d3rvich.database
+package ru.d3rvich.core.data
 
-import org.junit.Assert.assertEquals
 import org.junit.Test
+
+import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
