+++ core/remote/src/test/java/ru/d3rvich/core/remote/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
-package ru.d3rvich.data
+package ru.d3rvich.core.remote
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
